`Protobuf` converter providers don't attempt to illegally cast nested parameterized types (#4429)
Motivation:

`RequestConverterFunctionProvider#createResponseConverterFunction`, `ResponseConverterFunctionProvider#createResponseConverterFunction` is called to check if a request/response can be handled for all available request/response types in annotated services.

The previous `Protobuf[Request|Response]ConverterFunctionProvider` assumes that only 1-depth parameterized requests will be received. However, it is fairly common for json request/response objects to be defined with multiple levels of parameterized types.

This pull request fixes the current implementation so that at least a `ClassCastException` will not be thrown.
As a result of this PR, `Protobuf[Request|Response]ConverterFunctionProvider` will not handle nested protobuf types.
When used by default configuration, this causes a 500 while attempting to serialize/deserialize protobuf objects using jackson (json). e.g. the following return type will return a 500
```
List<List<Message>>
```
This is a separate point that will be addressed in the future.

Modifications:

- Check `ClassCastException` for `Protobuf[Request|Response]ConverterFunctionProvider` and it's scala counterparts.

Result:

- Users can now use `Protobuf[Request|Response]ConverterFunctionProvider` without encountering a potential `ClassCastException`.