Use `HttpHeadersBuilder` for additional headers in `RequestContext`
Motivation:

`RequestContext`s internally reference an `HttpHeaders` to hold
additionals headers which override the original headers set when
creating a request or a response.

`HttpHeaders` is an immutable data so it should be converted into an
`HttpHeadersBuilder` to mutate the values. In addition, atomic
operations are required to update the newly created `HttpHeaders` to
`RequestContext`.
https://github.com/line/armeria/blob/529738c836c78bf94d032f1328ec4241d662d19c/core/src/main/java/com/linecorp/armeria/internal/client/DefaultClientRequestContext.java#L722

If `RequestContext`s have a `HttpHeadersBuilder` reference internally,
1) we don't need to use atomic operations and 2) convert `HttpHeaders`
to `HttpHeadersBuilder` and back to `HttpHeaders`.

Modifications:

- Use `HttpHeadersBuilder` for `additionalRequestHeaders` in
  `DefaultClientRequestContext`.
- Use `HttpHeadersBuilder` for `additionalResponseHeaders` and
  `additionalResponseTrailers` in `DefaultServiceRequestContext`.

Result:

Simplify updating additional headers.

Note:

This PR does not seem to be listed in the release notes.