Return cached aggregated message for following withPooledObjects call (#4449)
Motivation:
We provide a way to cache `HttpRequest.aggregate()` in #4366. It can be mainly used in a decorator to get the body of an `HttpRequest`. When `HttpRequest.aggregate()` is called in a decorator, if the destination service is an annotated service, there's no problem. However, if a gRPC service or a Thrift service calls `HttpRequest.aggregateWithPooledObjects()`, the service raises an exception.

Modifications:
- Add `AggregationOptionsBuilder.usePooledObjects(alloc, preferCached)`
  - If the `preferCached` is true, it returns the cached aggregated message.

Result:
- `GrpcService` and `ThriftService` no longer throw double subscription exceptions when the request is aggregated in a decorator.
- (Deprecated) `HttpRequest.aggregateWithPooledObjects()` and `HttpResponse.aggregateWithPooledObjects()` are deprecated.
  - Use `HttpRequest.aggregate(aggregateOptions)` with `AggregationOptions#usePooledObjects(ByteBufAllocator)`.