Fix a bug where a connection is closed before a response is written in HTTP/1.x
Motivation:

If `connection: close` header is recevied, `HttpServerHandler` closes a
connection when an `HttpResponse` is completed.
https://github.com/line/armeria/blob/f6c4debed749f77d71aec836e98bdffef6537273/core/src/main/java/com/linecorp/armeria/server/HttpServerHandler.java#L412-L425
However, `HttpResponse` completion does not guarantee that the response
is fully written to the connection. If a response write starts after the
completion of `HttpResponse.whenComplete()`, a response could write to
a closed connection which causes
`io.netty.channel.StacklessClosedChannelException`.

Modifications:

- Start post-process of a response after the response is really written
  to the channel.
- Add `completionFuture` and make it complete when
  `AbstractHttpResponseHandler` finishes all write operations.
- Immediately close `responseEncoder` so that a late response is completed
  with a `ClosedSessionException`.

Result:

You no longer see `StacklessClosedChannelException` when handling
an HTTP1 request with 'Connection: close`.