Fix `ConcurrentModificationException` in `HttpServerHandler` (#4524)
Motivation:

If a connection is closed, unfinished requests are aborted.
https://github.com/line/armeria/blob/e94bbd075b745098a76e514e870a6c1803cbeec8/core/src/main/java/com/linecorp/armeria/server/HttpServerHandler.java#L236-L246
When the unfinished requests are aborted, their responses are aborted as well. The abortion signal completes `resWriteFuture` exceptionally and its callback tries to remove the request from `unfinishedRequests`.
https://github.com/line/armeria/blob/e94bbd075b745098a76e514e870a6c1803cbeec8/core/src/main/java/com/linecorp/armeria/server/HttpServerHandler.java#L418-L430
This recursive removal while iterating `unfinishedRequests` causes `ConcurrentModificationException`.

Modifications:

- Do not remove a request from `unfinishedRequests` if it is called in `cleanup()`

Result:

You no longer see `ConcurrentModificationException` when a connection is closed in a server.