Ensure the post-process of a response is executed without a local request context
Motivation:

A listener of a `ChannelFuture` can have a different request context in
the thread local because Netty may handle some pending requests in a
batch task which can may be triggered by a event loop having the
different request context.

Modifications:

- Remove a request context from thread local before invoking
  `tryComplete()`

Result:

Fixed a potential context leak when an incomplete `HttpResponse` is
returned.