Fix `FallbackService` not to raise `HttpStatusException`
Motivation:

We decided not to use `HttpStatusException` and `HttpResponseException`
internally because if an exception is thrown, it break through decorator
chains and we can't use stream operatorators in `HttpResponse`. #4117
Thad said, `FallbackService` still throws defered `HttpStatusException`s
which we should've migrated in #4117 but omitted it.

Modification:

- Return a normal `HttpResponse` instead of throwing
  `HttpStatusException` in `FallbackService`.

Result:

You no longer see `HttpStatusException` when there is no matching route.