Fix a bug where an EndpointGroup is not copied to a derived client.
Motivation:

When a new client is derived from an existing client using
`Clients.newDerivedClient()`, all options should be copied to the new
client and the original `EndpointGroup` should be reserved.
It seems to have already been considered by `newDerivedBuilder()` but there is one API missing.
https://github.com/line/armeria/blob/81ac954d41ee8420898ba298f9d0a766c3d37b71/core/src/main/java/com/linecorp/armeria/client/Clients.java#L317-L319
The following API called `builder` instead of `newDerivedBuilder`.
https://github.com/line/armeria/blob/81ac954d41ee8420898ba298f9d0a766c3d37b71/core/src/main/java/com/linecorp/armeria/client/Clients.java#L303-L306
As a consequence, a derived client try to send a requst to `armeria-group-`.
https://github.com/line/armeria/blob/81ac954d41ee8420898ba298f9d0a766c3d37b71/core/src/main/java/com/linecorp/armeria/client/DefaultClientBuilderParams.java#L92-L92

Motifications:

- Fixed `newDerivedClient(T, Function)` to call `newDerivedBuilder()`

Result:

`EndpointGroup` is now correctly copied to a newly dervied client.