Add Bytes that represents binary data (#4505)
Motivation:
We create an `HttpData` by wrapping `ByteBuf` or `byte[]` that represent binary data. `ByteArrayHttpData` or `ByteBufHttpData` is internally used to operate on the binary data. Because it's an `HttpData` we cannot share the logic if we add other classes that are binary data such as WebSocket frames.

Modifications:
- Add `Bytes` and its implementations: `ByteArrayBytes`, `ByteBufBytes`
  - `HttpData` uses them internally.
- Add more methods to `StreamDecoderInput` that will be used for WebSocket.

Result:
- Ready to work on #3904