Limit the max length of decoded content in `Decoding{Service,Client}`
Motivation:

`ServerBuilder.maxRequestLength()` and
`ClientBuilder.maxResponseLength()` can limit the max content at the
session layer.
https://github.com/line/armeria/blob/ec7c4251994f4b3b0089f95904dc1a68e5e9ce30/core/src/main/java/com/linecorp/armeria/server/ServerBuilder.java#L1657-L1662
The length-limited compressed content could be a humongous object if the
original content was compressed with a high ratio. Such big objects may
cause `OutOfMemoryError` even if the content length is limited at the
session layer. For that reason, it is necessary to check the length of
an output content while decompressing the compressed content.

Additionally, this PR also fixes #4469.

Note:

`ZlibDecoder` is able to limit the size at the decoder level.
Unfortunately, there is no such option `BrotliDecoder`, but an output
data is chunked into 4MiB if it exceeds 4MiB. We can't sophicately
enforce the size of output but 4MiB seems a acceptable size.

Modifications:

- Added a new API that takes the max length of a decompressed object to
  `StreamDecoderFatory`.
- Fixed `StreamDecoder` to raise `ContentTooLargeException` if the total
  length of an object exceeds a certain limitation.
  - Fixed `ZlibStreamDecoder` to hand over the given `maxLength` to
    `ZlibDecoder`.
  - Fixed `AbstractStreamDecoder` to limit the content length after
    decoding.
- Added `com.aayushatharva.brotli4j:native-osx-aarch64` as an optional
  dependency.
- Removed obsolete `client.encoding.StreamDecoderFactories` and
  `client.encoding.ZlibStreamDecoder` which can be replaced with the
  same classes in `common.encoding`.
- Made `ContentTooLargeExeption` take an cause which is most likely a
  `DecompressionException` raised by `ZlibDecoder`.
- Fixed `DefaultServerErrorHandler` to return `413 Request Entity Too
  Large` for `ContentTooLargeExeption`.
- Replaced `ByteArrayOutputStream` with `ByteBufOutputStream` in
 `HttpEncodedResponse` for zero-copy.
- A `Content-Encoding` is removed by `DecodingClient` and
  `DecodingService` when a compressed content is to be decompressed.

Result:

- `DecodingClient` and `DecodingService` can now limit the maximum length
  of decompressed data.
- Fixes #4469