Remove uncompressed JavaScript files from the core artifact. (#4568)
Motivation:

The size of an uncompressed js file for `DocService` is about 2.5MB. The size is quite huge. It would be worth compressing the file since modern browsers can decompress the file after receiving it. If some user agents do not support compression, DocService automatically uncompresses the content on the fly before sending out the file.

Modifications:

- Add `deleteOriginalAssets` option to `CompressionWebpackPlugin`
  https://webpack.js.org/plugins/compression-webpack-plugin/#deleteoriginalassets
- Add a unit test to ensure the original asserts are deleted

Result:

- Deleted uncompressed `DocService` assets that were added unintentionally.
- Fixes #4500