Fix a deadlock when a late response are written first in HTTP/1 pipelining (#4536)
Motivation:

Let's say a late request is completed first while the first request is being processed. The response to the late request is added to `pendingWritesMap` to preserve the order of responses in HTTP/1. Before the response is added to a queue, `lastResponseHeadersId` is updated to the ID of the late request.
https://github.com/line/armeria/blob/b127cd27252f6a454130f66d1175a06faed01f09/core/src/main/java/com/linecorp/armeria/server/ServerHttp1ObjectEncoder.java#L79-L95

It is a bug because the actual headers are queued and not written. Although a normal response for the first request is completed, the response is canceled because `HttpResponseSubscriber` assumes the order of response is reversed.
https://github.com/line/armeria/blob/0930b184bf00f98692a2e0b3eb044410e78a8173/core/src/main/java/com/linecorp/armeria/server/HttpResponseSubscriber.java#L141-L147

Modifications:

- Fixed `Http1ObjectEncoder` to use `currentId` which is updated only when a response starts to write for `isResponseHeadersSent()`.
- Reset a connection when two headers are about to be written to a response in HTTP/1. It is possible when a request is failed by `Http1RequestDecoder` with a protocol violation.

Result:

- HTTP/1.1 responses are now correctly ordered even when HTTP pipelining is enabled.
- Fixes #4470