Add useParameterAsRoot when creating MethodInfo (#4592)
Motivation:
Related: https://github.com/line/armeria/pull/4518#discussion_r1051738777

The layout of JSON requests for gRPC and Thrift are different. gRPC uses the method parameter as root but Thrift uses the method parameters with the key. For example:
```
// When struct is:
MyStruct {
  String name;
  int age;
}

// When the method is `myMethod(MyStruct myStruct)`
// gRPC:
{
  "name": "foo",
  "age": 40
}

// Thrift:
{
  "myStruct": {
    "name": "foo",
    "age": 40
  }
}
```
This also applies for JSON schema as well.
Please note that gRPC HTTP endpoints aren't taken account into because we create parameters in a `MethodInfo` with the struct properties.

Modification:
- Add `useParameterAsRoot` when creating gRPC `MethodInfo`.

Result:
- Resolve https://github.com/line/armeria/pull/4518#discussion_r1051738777