Generate JSON Schema from ServiceSpecifications (#4518)
## Motivation:

In order to auto-completion in debug form, we need to define JSON schema for the request body type. This is a relatively simple task because we already have a well-structured `ServiceInfo` object that is served with the doc service. This PR adds an expansion to the doc service server to also generate & serve JSON schemas.

- https://json-schema.org/learn/getting-started-step-by-step.html
- https://json-schema.org/understanding-json-schema/structuring.html

## Modifications:

- Add a new class to generate `JSON Schema` from `ServiceSpecifications`.
- Update `DocService` to serve JSON Schemas under the `/docs/schemas.json` endpoint as an array of objects.

## Result:

Example `schema.json` file for the `TestMessage` from `message.proto` which contains almost everything: