Fix a bug where `responseCause` is set to `ClosedSessionException` for a success response.
Motivation:

It would be legitimate that a HTTP/1 client closes a connection after
data is fully received. Even though, the connection is closed, the
stream on the server-side could be still open.
Because `HttpResponseSubscriber` may be complete after fully sending
data and then receiving `onComplete()` signal.
Receiving `onComplete()`, `HttpResponseSubscriber` tries to write
an empty chunk as a mark of EOS to the disconnected channel.
https://github.com/line/armeria/blob/fee87f8da942d0e6343e814489524b73db60ef3b/core/src/main/java/com/linecorp/armeria/server/HttpResponseSubscriber.java#L317-L320

Modifications:

- Check a connection is active as well when `Http1ObjectEncoder`
  determines a session is writable.
- Do not send EOS and mark a request as success when a HTTP/1 session is
  inactive when handling `onComplete()` in `HttpResponseSubscriber`

Result:

You no longer see `ClosedSessionException` when a connection is closed
after a response data has been fully sent in HTTP/1.