Fix a bug where a DNS resolution is timed out when IPv6 is enabled
Motivation:

If IPv6 is enabled, the default DNS resovler send both `A` and `AAAA`
record type to get available IP addresses. `DefaultDnsResolver` waits
for all questions to be resolved.
https://github.com/line/armeria/blob/94a1bf40ff41e1cfa9cd0c1a08f7bde14a7153c9/core/src/main/java/com/linecorp/armeria/internal/client/dns/DefaultDnsResolver.java#L126-L131

It is common that IPv6 record is not registerd for servers which is used
for internal communitation. If a DNS server does not have an IPv6
address for a DNS query, the default DNS resolver continues to look up
the next search domains.

Even if a DNS resolver acquired an IPv4 address successfully,
`DefaultDnsResolver` keeps waiting for an IPv6 response.
It would be a long running task to check all search domains.

For that case, users frequently see `DnsTimeoutException` without
`-Djava.net.preferIPv4Stack=true` which sends a DNS question only for
IPv4.

Modifications:

- Fix `DefaultDnsResolver` to make a DNS query success immediately if
  the first preferred records are found.
  - `DefaultDnsResoluver` does not wait for all questions are resolved.
  - If a less preferred response is recevied early, the resolver will
    wait for the highly preferred one until the query is timed out.
  - If resolved records exists when a timeout is triggered, use the
    value.

Result:

You no longer see `DnsTimeoutException` when IPv6 network inteface is
enabled.