Automatically delete branch if `Update public suffix list` PR is merged
Motivation:

According to the documentaion of [create-pull-request](https://github.com/marketplace/actions/create-pull-request#action-inputs),
a branch has to be removed if an automatically generated PR is merged.
I don't know why but it does not work as it is configured.

Mergify also provides an option that deletes branch when a PR is merged.
We can try the option as a workaround.

Modifications:

- Add a new step that removes the head branch of a pull request.
  - This option is applied to the autmotically generated PR.

Result:

Automatically delete `update-public-suffixes` after merging.