Fix javadocs for `WeightedRoundRobinSelector` (#4747)
Motivation:

Typo in `WeightedRoundRobinSelector` javadoc 

Modifications:

Below javadoc should be modified. It seems that the weights are 3, 5, 7 but the javadoc indicates as if `b` has the highest weight. 

<as-is>
`if endpoint weights are 3,5,7, then select result is abcabcabcbcbcbb abcabcabcbcbcbb ...`

<to-be>
`if endpoint weights are 3,5,7, then select result is abcabcabcbcbccc abcabcabcbcbccc ...`

Result:

Fix typo