Fix buffer leak with HEAD requests
Motivation:

If a HEAD request is received, a server should disgard the content of
a response and return headers only.
There was a regression that pooled data is not released if the response
body has more than two chunks.
https://github.com/line/armeria/blob/a83388c18b31d400f36ff7346902b30f9c194ee7/core/src/main/java/com/linecorp/armeria/server/HttpResponseSubscriber.java#L152-L161
https://github.com/line/armeria/pull/3956/files#diff-368938fb6a538f212fadf2ca09779cd5a6e05002c393454e7fc4b03361b588ccR153

Modifications:

- Cancel an `HttpResponse`'s body if HEAD method is received or data is
  published or a stream is closed.

Result:

Fixed a memory leak where a response to a HEAD request is not cleaned up
properly.