Introduce internal headers for client side (#4697)
Motivation:

Currently, headers are applied in the following order:

```
additionalHeaders > requestHeaders > defaultHeaders
```

This works fine for normal headers, but authority/host headers are unique in that `requestHeaders.authority` is overwritten when `DefaultClientRequestContext#init(EndpointGroup)` is called.
This can be problematic if users want to specify custom authority headers in default/request headers. Additionally, requests involving derived contexts are prone to mistakes. For instance, retried requests without setting the authority explicitly may send the wrong authority.

I believe the root cause of this issue is that armeria overrides headers that might have been set by the user.
For this reason I propose that we introduce a new `internalHeaders` that are set only by armeria internally. These headers would be at the bottom of the hierarchy under `defaultHeaders`.

```
additionalHeaders > requestHeaders > defaultHeaders > internalHeaders
```

Whenever `DefaultClientRequestContext` updates the target endpoint, `internalHeaders.authority` is also updated to reflect the new endpoint.
- If a user doesn't specify any authority headers, the endpoint authority will be used
- Otherwise, the user-specified authority header will be used

Note that this change means that `HttpRequest#headers` now represents the headers set for the request, and doesn't necessarily represent the entire set of headers that will reach the remote endpoint.
For this reason, users who would like to determine the authority/uri for a request should now refer to `ClientRequestContext#authority`, `ClientRequestContext#uri` instead.

**Note** The reason for using a backup headers is simply because I thought it is mentally easier to reason about. I think it's also fine to use `endpoint#authority` instead of `internalHeaders` if performance is really a concern.

Modifications:

- `BraveClient` and `CookieClient` now uses `ClientRequestContext#authority`, `ClientRequestContext#uri`.
- Introduced `internalHeaders` to `DefaultClientRequestContext` which is used last when computing the final headers to be sent.
- Introduced `DefaultClientRequestContext#authority` which returns the authority for a given request.
- Introduced `ClientRequestContext#uri` which returns the uri for a given request.
- `HttpHeadersBase#uri` now doesn't require `authority`. The reasoning for this change is similar to the behavior of `ClientRequestContext#uri`.
-  `ClientRequestContext#updateRequest` now also updates the `path` and `query` for implementing `ClientRequestContext#uri`. Note that `endpoint` and `sessionProtocol` are not updated and follow values from `initEndpoint`.
- `THttpClientDelegate` now doesn't set the `authority` header, and relies on autofill behavior.

Result:

- The correct host is set for retried requests