Deprecate `ClientFactory.setMeterRegistry()`
Motivation:

`ClientFactory.setMeterRegistry()` was first introduced as a way to
allow a user to override an already-instantiated `ClientFactory`'s
`MeterRegistry`. It was needed at that time to provide a way to change
the default `ClientFactory`'s `MeterRegistry`.

However, we don't need such a hack anymore because a user can override
the default `MeterRegistry` using `FlagsProvider` SPI.

Modifications:

- Deprecated `ClientFactory.setMeterRegistry()`.

Result:

- (Deprecation) `ClientFactory.setMeterRegistry()` has been deprecated
  in favor of `FlagsProvider` SPI.