Fix a wrong request path in `CustomServerErrorHandlerTest`
Motivation:

The test service is bound to `/bar` but the client sends a request to
`/foo`.
https://github.com/line/armeria/blob/b16bc8d926e3b4ebb07b86b66a09cd1126121574/core/src/test/java/com/linecorp/armeria/server/CustomServerErrorHandlerTest.java#L78-L82
https://github.com/line/armeria/blob/b16bc8d926e3b4ebb07b86b66a09cd1126121574/core/src/test/java/com/linecorp/armeria/server/CustomServerErrorHandlerTest.java#L150-L154
I believe that is a typo. The test ensures a exception raised in a
service is handled by an error hander at the virtual host.

The test has passed before #4552 is merged. Because `FallbackService`
raised `HttpStatusException` that is recovered to 400 Bad Request.

Modifications:

- Fix `CustomServerErrorHandlerTest.defaultVirtualHost()` to send a
  requset to `/bar`.
  - `AnticipatedException` is used to ensure the error handler
    catches the exception raised by the test service.

Result:

Fix CI build failures.