Add `AbstractRuleBuilder.onGrpcTrailers()` (#4535)
Related issue: #4496

Motivation:

A user might want to write a circuit breaker or retry rule that makes a decision based on gRPC trailers. Accessing gRPC trailers can be tricky because they can be present in either HTTP headers, trailers or even content (if gRPC Web). It would save a lot of duplicate effort to write a filter that checks difference places if we provide a built-in filter for gRPC trailers.

Modifications:

- Add `AbstractRuleBuilder.onGrpcTrailers()`
- Modify `AbstractRuleBuilderUtil` so it handles a gRPC trailers filter.
- Move the core logic of `GrpcWebTrailers` to `InternalGrpcWebTrailers` in the `core` module, because otherwise `AbstractRuleBuilderUtil` cannot access gRPC Web trailers.
- Add `AbstractRuleBuilder.requiresResponseTrailers()` to reduce duplication
- Add `AbstractRuleBuilder.combinePredicates()` to reduce duplication.
- Fix some Javadoc errors.

Result:

- A user can write a circuit breaker or retry rule that makes a decision based on gRPC trailers, regardless of whether it's a gRPC response with/without body or even gRPC Web response.
- Partially implemented #4496 although a little bit different from the initially suggested idea