Fix to use a relative path for schemas.json in DocServcie client
Motivation:

A `DocService` can be bound to any path although `/docs` is a common
path pattern in Armeria examples.
However, `/docs/schemas.json` is harded coded to fetch `schemas.json`.
https://github.com/line/armeria/blob/8652e5ea544a99790cab391274dde22370b96d35/docs-client/src/containers/App/index.tsx#L481
The resource will not be found on the debug page if users bind
`DocService` on `/internal/docs`,

Modifications:

- Use a relative path without a prefix.

Result:

Fix autocompletion not working when `DocService` is bound to a path
other than `/docs`.