Fixed a bug where aliases were erased from StructInfo
Background:

gRPC creates `StructInfo` from protobuf's `Descriptor`. As the generated
`Message` class is unknown in the `Descriptor`, only the message name
defined in the `.proto` file can be extracted to `StructInfo`.

However, annotated services create `StructInfo` from `Message` generated
by the protobuf compiler. The class name of the message is injected as an
alias into the `Structinfo` created based on the `Descriptor` so that
`DocService` links the parameter class to the protobuf `Structinfo`.

https://github.com/line/armeria/blob/8652e5ea544a99790cab391274dde22370b96d35/docs-client/src/lib/specification.tsx#L117-L123

Motivation:

If a generated protobuf `Message` is used with both a gRPC service and
an annotated services, `StructInfo.alias()` could be removed while
merging `StructInfo`s. Because `ServiceSpecification.merge()` dedups
`StructInfo`s that have the same name.

https://github.com/line/armeria/blob/6b7b1b0b638f3d780f5910a133daf24902af8d93/core/src/main/java/com/linecorp/armeria/server/docs/ServiceSpecification.java#L151-L155

Modifications:

- If two `StructInfo`s have the same name, preservce the one with the
  alias.
- Fixed `JsonSchemaGenerator` so that `StructInfo. can also be found
  by an alias.

Result:

`DocService` now correctly makes the link to `StructInfo` even when
a protobuf `Message` is used in both gRPC services and annotated
services.