Give high priority to the innermost `MetricCollectingService`
Motivation:

If duplicate `MetricCollectingService`s are used to decorate a service,
the first encounted `MetricCollectingService` is picked to collect
metrics.
https://github.com/line/armeria/blob/0df139a41cac430b3805f2774ac6d0c8ca4b4a68/core/src/main/java/com/linecorp/armeria/internal/common/metric/RequestMetricSupport.java#L57-L59

This strategy has one downside that users can't override a globally used
`MetricCollectingService` with a local decorator that is directly
applied to a service.

If we give high priority to the `MetricCollectingService` closest to a
service, users will be able to easily override the global one.

Modifications:

- Fixed `MetricCollectingService` to look for an inner
  `MetricCollectingService` before setting up request metrics.
  - If an inner `MetricCollectingService` is found, delegate the
    decorator to collect metrics.
- Moved `RouteDecoratingService` to `internal` package.
- Added `RouteDecoratingService.as(ctx, serviceClass)` to find a
  matching decorator from the internal service chain.

Result:

- The innermost `MetricCollectingService` is now choosen to measure
  metrics for a service.
  - Previously, the outermost one is used to collect metrics.