Preseve `%3F` (?) character in a request path (#4907)
Motivation:

`?` is not in the preserved encoding character set. As a result, if `%3F` or `%3f` is included in a request path, it is converted into `?`.
https://github.com/line/armeria/blob/0df139a41cac430b3805f2774ac6d0c8ca4b4a68/core/src/main/java/com/linecorp/armeria/internal/common/DefaultRequestTarget.java#L60-L60
`/foo/x%3Fy.json`, for example, is decoded into `/foo/x?y.json`. When a server parses the request path, `/foo/x` will be the path component, and `y.json` will be the query component. Although they wanted to send a path ending with `.json`.

Modifications:

- Add `?` to `PATH_MUST_PRESERVE_ENCODING` so that `RequestTarget` does not decode `%3F` in a path into `?`

Result:

You can now correctly send percent-encoded `?` in a request path.