Fix a bug where `RequestLog.responseCause()` isn't recorded for RPC responses
Motivation:

Recently, i've encountered the following test failure repeatedly when I
was running tests locally.
```
GrpcClientTest > errorNoMessage() FAILED
    java.lang.AssertionError:
    Expecting actual not to be null
        at com.linecorp.armeria.client.grpc.GrpcClientTest.lambda$errorNoMessage$44(GrpcClientTest.java:1639)
        at com.linecorp.armeria.client.grpc.GrpcClientTest.checkRequestLogError(GrpcClientTest.java:1863)
        at com.linecorp.armeria.client.grpc.GrpcClientTest.errorNoMessage(GrpcClientTest.java:1634)
```

That regression was caused by the change in #3904 that added
`responseCause(Throwable)`. The PR also migrated the way a response
cause is set from `RpcResponse`
https://github.com/line/armeria/pull/3904/files#diff-91096a8a0f87541eb2debf804754395c3000432ccc7af9167fd99d93e4241628

`RpcResponse` could be set after an HTTP response is ended.
This is because the decode process takes place at the RPC layer.

Note:

This regression has not yet been released, so there is no impact on
users.

Modifications:

- Direcly set `RpcResponse.cause()` to `responseCause` if it is null.

Result:

`RpcResponse.cause()` is correctly propagated to
`RequestLog.responseCause()`.