Map `FailFastException` to gRPC `Status.UNAVAILABLE`
Motivation:

`FailFastException` is mapped to `Status.UNKNOWN` while
`UnprocessedRequestException` is mapped to `Status.UNAVAILABLE`.
https://github.com/line/armeria/blob/4c9dccf996dc062f8f77d7013de504141d9c8306/grpc/src/main/java/com/linecorp/armeria/internal/common/grpc/GrpcStatus.java#LL125C1-L125C1
Both `FailFastException` and `UnprocessedRequestException` mean that
a request did not work on a server, so it would be better to map to the
same `Status`.

See #4805 for the detail.

Modifications:

- Fixed `GrpcStatus.fromThrowable()` to convert `FailFastException` to
  `Status.UNAVAILABLE`

Result:

`FailFastException` is now correctly converted to `Status.UNAVAILABLE`.