Fix a bug where "connection: close" header is stripped  in a HTTP/1.1 response (#4864)
Motivation:

`connection: close` header that should be preserved is removed while an
Armeria server response is converted to a Netty HTTP/1 response. We
allowed users to send `connection: close` #4531. The connection was
correctly closed after fully writing the response but `connection:
close` wasn't sent correctly.

Modifications:

- Make `KeepAliverHandler` closed when `connection: close` header is
specified in the response headers.

Result:

In HTTP/1.1, if `connection: close` is set in response headers, the
value is correctly sent to clients.