Fix a bug where content preview is not logged
Motivation:

I got a report that a content preview is not logged with
`Logging{Service,Client}`.

Previsouly, a content preview is logged when a sanitized content is null.
https://github.com/line/armeria/blob/f6f479de876456aca34e231d2f7071c2fa26fd5d/core/src/main/java/com/linecorp/armeria/common/logging/DefaultRequestLog.java#L1528-L1530
When we migrated the code to `LogFomatter`, the condition has changed
slightly. The new condition did not attempt to read a content preview
when `RESPONSE_CONTENT` is available.
https://github.com/line/armeria/blob/f221c8fda5b92412d9985ea659d085168835aa05/core/src/main/java/com/linecorp/armeria/common/logging/TextLogFormatter.java#L229-L234

`RESPONSE_CONTENT` flag can be set with a null value for REST API. If
the actual content is null, we need to check whether
`RESPONSE_CONTENT_PREVIEW` is availble.

Modifications:

- If the value of `REQUEST_CONTENT` or `RESPONSE_CONTENT` is null, try
  to get content preview.

Result:

You can now correctly see a content preview from `LoggingService` and
`LoggingClient`.