Propagate TLS handshake exception in `SslCompletionEvent` (#4950)
Motivation:

- `SslCompletionEvent` can have a cause if a TLS handshake is completed
exceptionally. As the cause isn't taken into account, users could see
the root cause of the failure.

- A legacy HTTPS server such as Microsoft-IIS/8.5 resets the connection
if no cipher suites in common without respecting the handshake process.
No hint is given to the handshake failure, users don't know why Armeria
can't communicate with the server.

Modifications:

- If `SslCompletionEvent` is not successful, propagate the cause to the
session promise.
- If an exception is raised during a handshake, create
`IllegalStateException` with a hint and add it as a suppressed
exception.

Result:

Exceptions that occurred during a TLS handshake are properly propagated
to users.