Fix a bug that default logger is changed for LoggingService and LoggingClient (#5056)
Motivation
I changed the default logger of `LoggingService` from `com.linecorp.armeria.server.logging.LoggingService` to `com.linecorp.armeria.common.logging.DefaultLogWriter` in #4943. This caused a breaking change to logging system if you are using the class name with a logger. This also applies to the `LoggingClient`.
We should fix it.

Modification
- Reverted to use `com.linecorp.armeria.server.logging.LoggingService` logger for `LoggingService` when `LogWriter` isn't set.
- Reverted to use `com.linecorp.armeria.client.logging.LoggingClient` logger for `LoggingClient` when `LogWriter` isn't set.

Result
- As in versions prior to 1.24.0, the default logger for `LoggingService` is now `com.linecorp.armeria.server.logging.LoggingService`.
  - The default logger for `LoggingClient` is also `com.linecorp.armeria.client.logging.LoggingClient`.