Make `RequestScopedMdc` compatible with Logback 1.4.8 (#5079)
Motivation:

Logback 1.4.8 no longer uses `MDC.getMDCAdapter()` to get `MDCAdapter`.
Instead, `MDCAdapter` is acquired from `loggingContext` set to
`LoggingEvent` at construction time.

`RequestScopedMdc` replaces the underlying `MDCAdapter` of `MDC` using
reflection (a hacky way) to inject MDC stored in `RequestContext`. Due
to the change in Logback's implementation of `MDCAdapter`, the replaced
`MDCAdapter` can not be integrated with Logback's MDC.

AS there is no way to override the default `LogbackMDCAdapter` with a
public API, I have to use another reflection to inject a custom
`LogbackMDCAdapter` to `LoggingContext`. It would be worth making an API
request to Logback to avoid unexpected behaviors because of internal
changes in Logback.

See: #5045 

Modifications:

- Created `DelegatingLogbackMDCAdapter` that is injected to
`LoggerContext` to replace the default Logback's `MDCAdater`.
- `DelegatingLogbackMDCAdapter` is only effective for Logback 1.4.8 or
above.
- Logback 1.4.x is added as a compile dependency to the `core` module.
- Bump Logback version 1.4.8 from 1.4.7.
- Create `:it:logback1.4` module to test the compatibility between
`RequestScopedMdc` and Logback 1.4.8.

Result:

- `RequestScopedMdc` is now compatible with Logback 1.4.8
- Fixes #5045